requirements.txt
--pre
--extra-index-url https://download.pytorch.org/whl/cpu

python-dotenv
fastapi[all]==0.115.6
uvicorn==0.32.1
openvino_genai==2025.1.0.0

basicsr@git+https://github.com/yanhernt/BasicSR
numpy==1.26.4
librosa==0.10.2
openvino==2025.1.0

setuptools
face_alignment==1.3.5
imageio-ffmpeg==0.4.7
numba
resampy==0.3.1
pydub==0.25.1 
kornia==0.6.8
tqdm
yacs==0.1.8
pyyaml  
joblib==1.2.0
facexlib==0.3.0
gradio
gfpgan
av
safetensors
gdown

# Serving
sse_starlette==2.1.2

# Utils
scipy
soundfile
openai==1.56.2
pypdf==5.0.0
langchain==0.3.8
langchain-chroma==0.1.4
langchain-community===0.3.8
chromadb==0.5.18
huggingface_hub>=0.23.0
botocore==1.34.88
cached_path

# OpenVINO
torch>=2.6.0
torchaudio>=2.6.0
optimum-intel[openvino,nncf]==1.22.0


torchvision
transformers==4.48.0
tyro==0.8.14
opencv-python==4.10.*
imageio==2.36.0
onnxruntime-gpu==1.20.0
onnx==1.17.0
requests==2.32.3
scikit-image==0.24.0
pykalman==0.9.7
imageio[ffmpeg]==2.36.0

#piper-tts
pydantic
piper-tts
piper-phonemize-cross
onnxruntime

hparams
lws